requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
click==6.7
Flask==1.0.2
gast==0.2.0
grpcio==1.11.0
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.6
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.3
Pillow==5.1.0
protobuf==3.5.2.post1
PyYAML==3.12
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
